+++ src/boot/boot_requests.c
@@ -45,6 +45,12 @@ volatile struct limine_executable_file_request exec_file_request = {
 };
 
 __attribute__((used, section(".limine_requests")))
+volatile struct limine_executable_cmdline_request exec_cmdline_request = {
+    .id = LIMINE_EXECUTABLE_CMDLINE_REQUEST_ID,
+    .revision = 0
+};
+
+__attribute__((used, section(".limine_requests")))
 volatile struct limine_executable_address_request exec_addr_request = {
     .id = LIMINE_EXECUTABLE_ADDRESS_REQUEST_ID,
     .revision = 0