+++ src/boot/boot_requests.h
@@ -11,6 +11,7 @@ extern volatile struct limine_memmap_request memmap_request;
 extern volatile struct limine_mp_request mp_request;
 extern volatile struct limine_bootloader_info_request bootloader_request;
 extern volatile struct limine_executable_file_request exec_file_request;
+extern volatile struct limine_executable_cmdline_request exec_cmdline_request;
 extern volatile struct limine_executable_address_request exec_addr_request;
 extern volatile struct limine_module_request module_request;
 extern volatile struct limine_rsdp_request rsdp_request;